fixed log book floating precision for now (passed price to ints by using cents) (temporary)

+++ MarketSimulator/Include/UI/Window/orderBookWindow.h
@@ -11,7 +11,7 @@ struct SOrderRow
 {
     unsigned int orderId;
     EDirection side;
-    double price;
+    unsigned int price;
     unsigned int quantity;
 };
